BEN-230 First pass at implementing anchor finding and handling latent data
This diverges from the sequel implementation in some respects.  Since
we have more zmw per lane, and a shorter lookback, we will uniformly
trace back anchor finding the maximal distance.  Sequel did an entropy
computation that theoretically decreases monotonically as you move backwards.
I'm not sure how much of that was paranoia that things aren't really
monotonic, and how much was trying to find the latest anchor point possible
(to limit the size of latent data if entropy stops decreasing).  In
either case, for now at least, mongo is *not* doing the entropy computation,
though it may be added later for verification purposes.