BEN-230 Played around a bit with __launch_bounds__
This is a way to enforce limits on the compilers choice for
things like register usage.  Using fewer registers can increase
occupancy, but also will cause/increase register spills into
global memory.  Whether that is a win or not depends on your
problem, which is why these compiler hints exist