BEN-649 Added const semantics to "view" datatypes
"const" is now embedded in the template parameter itself
when appropriate.  This does provide robust and compiler
enforced const propagation for the underlying data exposed
via "view" types like "DeviceView", but it does mean the
constness of the underlying data is separate from the usual
cv-qualifiers on the variable name.  For example if you do
"const auto& view = " the constness of the payload is hidden
in the `auto`, and has nothing to do with the `const` qualifier
on `view`.  Care may have to be taken to avoid writing code
that is confusing for a human reader