BEN-649 Fixed warnings about the compiler being unable to determine stack size in debug builds
The warning was (mostly) harmless, but quite annoying.  There was no real runtime
recursion, save for the fact that some compile-time recursive functions terminated
in such a fashion that runtime recursion was almost present, but prevented by
a conditional that always evaluated to false.  This meant that optimzed builds
managed to do away with things entirely, but debug builds could not see that and
were genuinely unable to determine the necessary stack size.  It's guesses were
(so far) sufficient, but the warnings were numerous and noisy.