BEN-649 Fixed bug that can cause a cudaErrorLaunchOutOfResources
DeviceOnlyArray has a generic constructor that runs on the gpu.  Some
invocations were having errors in debug mode, because the debug version
of the constructors required more registers than could be supported
with the hard coded 1024 threads per block in use.  Now it querries
the device and the code to dynamically select kernel execution
parameters that are maximally parallel, but do not excede the
physical hardware limits