BEN-816 Clean up minor hole in UnifiedCudaArray semantics.
Technically we were capable of not properly calling the destructors.
Now destruction is still cheap in the garden path, but if destruction
is non-trivial and the most recent copy of the data lives on the gpu,
we'll download it first.

This change does not affect any types in use now, just makes things
proper for future arbitrary types