BEN-816 Added missing synchronization points
In a number of places we had functions that accepted something like
a TraceBatch by value, called an asynchronous cuda kernel, and while
that was still running exit the function (thus destroying the
TraceBatch it was using).

In reality due to how memory pooling works and other things, the
pipeline was mostly functional, though in some cases different
cuda contexts were accidentally using the same scratch work space
and stomping on each other.

This is a short term resolution.  The new synchronization points are
not necessarily ideal, and we may want to take another strategy
to avoid destroying objects still in use by active kernels