BEN-893 Created basic wrapper for cuda kernel invocation
Right now it's just a proof of concept, that captures a
kernel function pointer and an arbitrary set of invocation
parameters, and managed the invocation itself.  What would
be `Kernel<<<launch, params>>>(function, params)` would now
be expressed as `PBLaunch(Kernel, launch, params)(function, params)`

The two sets of parenthesis are necessary because both the function
arguments as well as the kernel invocation parameters themselves,
can differ in length and type, so two variadic template packs are
required.