BEN-893 Implemented kernel tracking based off cuda events
The concept of a StreamManager has been introduced.  Data management
objects like UnifiedCudaArray now have one as a member, and when
the class is provided with a KernelLaunchInfo object (previously
just named LaunchInfo), the manager digests it.  The KernelLaunchInfo
contains information about the stream in use and the cuda event
associated with the kernel launch, and this information is now
used to detect errors such as deleting the data while still in
use on the gpu