BEN-1073 Moved a couple more variables to shared memory
I'm actually surprised this made things faster (roughly 3.5%), as the increased
shared memory usage further limits our maximal occupancy.  However for whatever
reason our actually achieved occupancy didn't drop along with the theoretical,
so this still ended up a win.  Leaving it for now, but I won't be entirely
surprised if this needs to be revisited again in the future