BEN-1073 Fixed minor bug in UnifiedCudaArray
The testCircularBuffer test was getting a non-const view of data slated for
upload to the gpu.  Since that data was marked as upload-only, getting that
host view does not cause the code to synchronize and wait for outstanding
kernels.  The error detection code then correctly determined that there was
a risk for the data to be modified before it is uploaded.  The test has been
updated to get a const view, and there is now no issue.

On a related note, the const accessor was missing an error check as well.
It needs to make sure a separate thread did not cause the download, which
would break the normal synchornization logic in UnifiedCudaArray