BEN-1073 Encapsulated bitpacking operations for labels into a PackedLabels class
This was somewhat of a delicate change.  One would expect that with PackedLabels
being simple, and it's functions easily inlineable, that this change would have
zero effect on the generated code.  However the first version of this change
caused the optimizer to make difference choices, and the packedLabels got pushed
out to global memory, with the added memory traffic killing performance (almost 2x
slowdown). Changing a template function to a generic lambda did seem to remove the
issue (and in fact sped things up slightly), though again with mandatory function
inlining, one would not expect the two to perform differently.