BEN-1141 Made UnifiedCudaArray and DeviceOnlyArray more threadsafe
The threading model of UnifiedCudaArray has been updated.  Previously
there has been a strict requirement of one host thread to one GPU stream,
but that is about to be relaxed to allow nested parallelism in host filters.
That required a little more locking to make sure multiple threads can
concurrently access the host data, while avoiding race conditions or
triggering multiple downloads.

It is still a requirement that only one thread hand UnifiedCudaArrays
to kernels, so that we can rely on the cuda scheduler for sensible
work ordering.  Multiple threads invoking kernels on the same data should
still be detected and trigger an error.

The expected usage of DeviceOnlyArray has not changed, though
there was a race condition where simultaneous access by multiple threads
could break the data structure used to keep track of which streams are
currently using the data.  This race condition has been fixed.