BEN-1589 Implement BlockRepacker
This is a fairly general implementation.  It requires that all incoming
packets have a uniform block size and that the block width is a multiple
of 32.  The packet frame count must evently divide into the batch frame
count.  Beyond that, anything goes.  The implementation supports
multiple threads, and also has a hook that allows the low-level implementation
to be replaced in special circumstances, if a less general purpose approach
ends up being faster in some cases.