BEN-1113 Updated quadratic accumulations in the BaselineFilter to use 32 bits
At a rough initial glance this caused about a ~5% slowdown in the
baseline filter.  But as that filter only comprises ~5% of the overall
GPU compute budget, it doesn't really affect the bottom line

For posterity sake, I've not profiled things in detail, but the
cause is likely because this added an extra 6 bytes per zmw to
the shared memory usage (out of a previous 24 bytes).  The 24 bytes
was enough to exactly fill the shared memory buffer with full occupancy,
so the additional memory usage may have come at a slight occuapncy
decrease (or maybe not, if something else was already limiting
occupancy)