PTSD-38 Achieved initial replacement of LaneArray.
The old LaneArray (and LaneArrayMask and LaneArrayRef) implementations
are fully ripped out. The previous implementation had two main
limitations:
 1. A LaneArray inherited from a LaneArrayRef, which meant it
    both had the array of data as well as a pointer.  This
    interacts poorly with data alignment and makes it not
    suitable for vectorization
 2. The set of operators/functions provided was not as flexible
    as those that exist for primitive types.  It meant a lot
    of explicit casts had to exist in code that would not
    be necessary for regular primitives.

The new version gets rid of the concept of a LaneArrayRef.  That
was used in order to treat arbitrary data as a LaneArray without
a copy.  However vectorization requires loading data into special
registers anyway, which is already pretty similar (if not exactly
the same) as a copy.  So now you just create a new LaneArray
upfront and the "copy" is the exact same as the loading of the
data into SIMD registers that would have to happen anyway.  After
that, if there are enough registers then all temporaries stay in
register, and if there are not enough registers, then the pushing
of the LaneArray data to the stack will be the same (or at least
similar to) the push/pop operations that would have to happen
even were there a LaneArrayRef type still.

The LaneArray idea is implemented in 4 basic parts:
1.  BaseArray: a CRTP base in charge of all construction/conversion/etc
    logic
2.  ArithmeticArray: An intermediate class that provides all the
    standard mathematical operations common to all types
3.  LaneMask: essentially an array of bools
4.  LaneArray: The array of primitives spanning a lane.