PTSD-38 Reworked the backend of the implementation a bit.
The overload set that exists for construction was simplified and
reduced, both making the code nicer and providing a decent
performance boost as will be explained below.  Beyond that
all operator overloads were moved to be inline friends, to
make sure they can only be found by ADL and at least one
of the arguments is a LaneArray type.

Now a bit of a back story:

The intel compiler was having some difficulties performing
"Return Value Optimization" in some cases.  When creating
a new LaneArray, a small chain of functions was set up to
handle some template logic, with the bottom of the stack
actually creating the return type.  The functions were
designed to be easily inlined and inspection of the assembly
verrified that they in fact were, but for whatever reason it
was completely failing to elide the temporaries created as
things were passed back up the stack.  Intel was literally
pushing and popping the same data to the stack multiple times
in multiple locations, and none of the intermediate loads
were ever used again.

So re-organizing things so variables were created at the top of the
call stack and passed down for modification sped things up
noticibly.