PTSD-207 Add Tests for histogramming module
There are two flavors of tests, correctness and performance.  The correctness tests
use simple constant or sawtooth inputs so that the expected bin values can be
known a priori.

The performance tests are more heavy-weight and come in a variety of flavors.  The
PicketFence signal generator is used, but we also explicitly mess with both
spatial and temporal correlations, as it is expected that these correlations
will affect performance for some implementation choices on GPU hardware. Spatial
correlations are handled by either have 1 or 64 unique signals per lane, and
temporal correlation is handled by either sing the picket fence signal as provided,
or sorting/randomizing it.