PTSD-147 Added the concept of a StashableDeviceAllocatin and a DeviceAllocationStash
The former is a device allocation that can, at need, be put into "cold storage"
on the host, allowing it's GPU memory to be re-used for other purposes.

The latter is a management class that can orchestrate the upload/download
of data for a given pool of interest.

The combination of the two will serve as a foundation for allowing
algorithm state data that uses more memory than the GPU contains, and
only keeping a select active portion on the GPU at any one momment