PTSD-200
Refactored how the TraceHistogramAccumulator works.  Previously it did two things,
both the binning of trace data as well as the conversion of baseline statistics to
histogram bounds.  Now it's been broken apart and the whole detection model estimation
bit has been reworked as follows:

* TraceHistogramAccumulator has been broken up, with things relating to baseline statistics
moved to a new SignalRangeEstimator class.  There remains a virtual dispatch in both cases
allowing a GPU/CPU dichotemy
* The old DetectionModelEstimator has been renamed to CoreDMEstimator
* A new DetectionModelEstimator has been introduced which handles the
  "buisiness logic" of the dme.  It mainly contains and orchestrates the use
   of the TraceHistogramAccumulator, SignalRangeEstimator and CoreDMEstimator.
   The new DetectionModelEstimator is no longer a virtual class and will have no
   GPU implementation.  It also will not look at any lane level detail, with
   actual number crunching reserved for the three constituent members mentioned
   previously.

As a side effect this simplifies some of the BatchAnalyzer code, as the logic for
coordinating the histograms and dme were moved from there to the new DetectionModelEstimator