Merged variance computation loop with the loops in the E-phase.
This required adding a correction term to account for using an
 old estimate for mu.  The cram tests pass, but the data there
seems nice enough to work even without the correction term, meaing
I may not have found subtle bugs.

This is allows the removal of the huge local arrays that scale
with the histogram size.  We can do the necessary calcualtions streaming
without saving data, which vastly decreases our load on the memory
subsystems