PTSD-200 Changed a few API functions to return by value instead of reference
Data being shared between filters is typically stored in a UnifiedCudaArray
to allow us to dynamically select mismatched GPU/CPU implementations. However
in this case the UnifiedCudaArrays were long-lived, which means if the downstream
filter was on the GPU, then after the analysis of a full chunk, each pool would
still be holding onto memory for that pool, which in the case of the histograms
at least is way too much.

Returning view copy instead of reference solves this problem, as the data
gets destroyed at the end of a batch's analysis.  This does force extra
data movement where it was not required before, but since this is happening
with a relatively slow frequency the added cost likely does not matter.  If it
does prove problematic in the future we'll have to do something else to counter
it, as is returning by reference will make the DME on the GPU use far too much
memory