PTSD-203 Updated code to use the new DataSourceBase API.
A minor update was the change to use the new Start function
exposed by DataSourceRunner, which needs to be called before
any data can be gathered.

The more major update was to provide definitions for the new
PacketLayouts function, describing how data is organized in
a chunk.  Using this information also necessitated adding a
virtual virtual interface to Repacker to transform
packetLayouts into BatchDimensions