PTSD-421 Fixed type conversion issue when invoking cuda kernels
When using a non-const container to launch a kernel that wanted
a const view, the code would fail to compile because that required
two user defined conversions, one from a DeviceHandle to a DeviceView,
and then a second from a DeviceView of non-const to a DeviceView of
const.  Now we can construct a DeviceView of const directly from
a DeviceHandle of non-const, which drops things down to a single
user defined conversion, which makes the code behave as you would
expect.