PTSD-421 Added FieldSerializers
These are the basic building blocks for serializing pieces of data to the baz file.
For all of them you specify the number of bits to use for the "main" portion,
and each has a different strategy for what to do when the supplied value does not
fit in the specified number of bits.  It can either:
  * Truncate (wraparound overflow).  Mainly for use when you know it won't overflow
    and don't want to pay for extra data checks
  * Simple Overflow: If the value will not fit in the specified number of bits,
    then those bits are all set to 1, and a fixed sized overflow value will be
    written subsequently
  * Compact Overflow: A variable length encoding where each both the main portion
    and each overflow byte has one bit reserved to indicate if there is a(nother)
    overflow byte still to be read.