PTSD-421 Added proof-of-concept serializer using switches/visits rather than compile time configurations
The code was done as a proof of concept since the ppa side deserialization will want
to be as general as possible.  We don't necessarily want to re-spin a PPA build and keeep
track of a million version breaks every time we tweak what information basewriter serializes.
While it's definitely slower than the template solution (by at least 6-7x), doing things this way
means we don't need to recompile PPA if we just decide to serialize a new field.  We only have
a version breakage if we change what transforms/serializations are supported.

The current POC is a bit of a mess, partially beacuse it's going all the way to/from the pulse
format.  PPA currently uses just a vector of vector of ints to store the data, and a
deserializer targeting that should be easier to write