PTSD-421 Fixed segmentation fault resulting from static PBConfig variable
This one was a bit of work to untangle, and I still don't fully understand everything.

The symptom was a use-after-free in ScopedProfiler code, used in the ManagedAllocations.cpp
code.  This intermittently caused a segfault, but the use-after-free happens very reliably
once you know how to look for it.  The confusing part is that the use-after-free can
only happen if a `static` variable is destroyed before a `static thread_local` variable,
which as far as I understand it should be impossible.

More confusingly, this issue only appears with the introduction of a static PBConfig
variable in the HostSimulatedPulseAccumulator class.  I do not see why the two issues
would be related. I did notice that if I tweak the PBConfig framework to eliminate
all usage of thread_local variables, the problem goes away.  I need to do further
work to see if this is a viable change to make in the PBConfig framework

For now, the static member of HostSimulatedPulseAccumulator has been changed to
be a unique_ptr of a config object.  Now the configuration object itself isn't
constructed until main as begun executing, and things seem reliably better now