PTSD-421 Worked around issue relating to incorrect IPD
The original code made a mistake, where the "IPD" was calculated as the
difference between two start frames, rather than from the end of one
pulse to the beginning of the other.  This has been fixed, but that
fix highlighted a problem with the current code, which I've currently
worked around, but will need to be addressed more robustly shortly.

The issue is when deserializing the baz data and going all the way back
to the orginal pulse structure (which stores start frame rather than ipd),
it's required that the pulse width is deserialized before the IPD.  I've
set up the tests to do this for now, but that is obviously beyond brittle,
and it will be fixed more robustly in the very near term.