PTSD-426 Added the concept of a GrowableArray
This is designed to handle holding ZMW data, where we have an upfront
guess of how much data things might require, but noisy ZMW might overrun
that initial guess and require more memory.

In spirit, this could almost be done with a standard container like
std::deque.  The main reason it isn't is so that we can more easily
plug into the IAllocator framework and some of the usage monitoring
utility that provides.