PTSD-426 Added PacketBuffer class for holding packet data
The general use case is that we'll be streaming in small intervals
of pulse data (e.g. 512 frames) to fill up a larger buffer (e.g. 4096
frames).  From the maximum pulse rate we have an idea of how much space
we will need, but it's only a guess and we might have to add additional
storage for any particular ZMW

This results in an ad-hoc linked list structure, which is a bit brittle,
but focuses more on space economy than anything else.  A single ZMW
will probalby only hold ~320 bytes of data, so early naive implementations
managed to get to a full 20% overhead with only a few extra pointers and
such running around.  The current implementation is more like 5% overhead