PTSD-687 Revisited the gpu baseliner implementation to use 8 bit data for the morphological filters as well
Currently the baseline filter is the largest user of persistent state data, at almost 9GB per 25M ZMW.  If we
convert the morphological filters to use 8 bitdata as well, that cuts things down by almost 4GB, which isn't
an earth shattering change to our PCIe IO budget, but is still a nice savings nonetheless.  Plus with each
thread now operating on a PBUChar4 (so that it can maintain 32 bit accesses), our baseline filter runtime
drops by ~20%.  This again isn't a huge change to the overal compute budget, but still something that is
nice to have

This is a crap commit, but things were updated to use 8 bit data in the state data as well.  This cuts out ~4GB of state data, and also makes the filter run ~30% faster.  This needs a fair bit of cleaning and robustifying however if it's going to go in