PTSD-538 TraceSaverBody now stores batchID and hole number
On a related note, TraceSaverBody was refactored a bit to create it's own TraceFile.
The TraceFile class was never really meant to be standalone, but rather a mid-level
building block.  The previous code was already starting to get confused with some
trace file data being populated at the site of construction, and other parts populated
by the TraceSaverBody.