PTSD-539 Fixed bug when generating a single baz file
The core issue is that a baz buffer was only generated for IO when the last batch
processed had metrics present, but this doesn't work when there are multiple
batches.  For one, the metrics are affected by the DME stagger schedule, so not
all batches provide metrics at the same time.  For another, the order in which
batches are processed is not deterministic.  It's possible to go through several
complete rounds of metrics, all while never having the last batch processed be
one of the batches producing metrics.

The result of this was that metrics end up being overwritten, since the buffers
currently only support a single metric per baz buffer.  This commit changes things
so that baz buffers are produced on a rigid and fixed schedule.  This fixes the race
condition and metrics are no longer clobbering each other.  This does leave an
issue where the pulses associated with a metric block don't line up perfectly in
the lookback buffers.  Each metric ends up associated with some pulses from after
the metric interval, which will affect preHQ finding a little bit.