PTSD-846 Updated cram test results
This was a bit messier than hoped.  Some of the changes are an expected
slight shift in values, that is expected to come from computing an hfmetric
block over a slightly different interval.  However some of the changes are
entirely new metric blocks inserted, ostensibly because the activity labels
have changed and so adjacent metrics could not be aggregated.

With a lot of help from Martin Smith, some of the changing activity labels
were looked at in more detail.  It appears we were already marginal in
some cases, and the slight shift in values caused us to switch some A0
labels to A1.  It was already perhaps a bit weird that they were A0 to begin
with seeing as these ZMW have a pulse rate of ~2.4

It's unclear how much of the activity label issue is from the nature of
this simulated data or if things need to be retrained for the Kestrel
world.  In either case, the cram test does appear to show expected behaviour
given the produced activity labels, and some ZMW do still show
metrics aggregation functioning as expected.  So at least for now we'll
just update these results, and hopefully in the future we'll either have
more robust testing data or a more robust analysis (or both)