PTSD-796 Added a basic edge frame classifier
Plugged it in as a standalone kernel.  For now edge frames get marked as the lowest signed short,
effectively putting them in the outlier bin, which isn't a long term solution.  The purpose of
this commit is to effectively set a performance baseline, against which a more integrated approach
can be compared.  For reference, the standalone filter takes ~2ms to run, which is about the duration
of the production binning implementation itself.