PTSd-363 Refactored how ManagedAllocations works
A large portion of the original motivation for this code was to provide
a global allocator that low-level implementations could use, keeping
the code as close to a vanilla new/delete as possible while still
re-using allocations for efficiency

However having a global allocator introduces lifetime issues, where if
you are not careful we'll be deallocating cuda allocations after the
cuda runtime itself is already gone, which causes a hard crash.  The
previous version worked around this by carefully controlling the order
of construction and destruction of static variables, to help guarantee
that the cuda runtime lasted longer than any of the allocation caches
that relied upon it.

An unfortunate side effect of this approach was that we do a trial
cuda allocation (and later a trial huge page allocation), which forced
us to rely on the presence of a GPU even if we were doing a pure host
run, which obviously is not desirable.  The whole goal of this refactor
is to remove that constraint.

The new approach sets things up so that the default allocation mode
doesn't cache things, which makes it safe during static teardown.
Changing the mode to either use caches or a new allocator (to use
pinned memory) now uses RAII to make sure that we restore things to
default by the time we exit scope.  This allows high level code
to set the allocation strategy for a low level implementation, without
running into any static lifetime issues during program exit.

Since the lifetime of the caches are already being changed, they
have been udpated so that individual IAllocators can have their
own cache or share a "global" one, where the "global" one only lives
as long as all of it's constituent IAllocators.  The memory utilization
reports have also been augmented to keep track of how much data is
in the cache itself, as well as how much the cache is helping or hurting
our overal memory usage