PTSD-599 Moved the FrameLabeler out of the prototypes directory
The FrameLabeler is one of our oldest modules, predating a lot
of the Kestrel architecture, and as such had some extra indirections/
shims in place to make it fit.  Some of those shims were getting in
the way of implementing the roi functionality, so I finally bit
the bullet and tried to fix things.  The code has been lightly
cleaned, the implementation moved out of the prototype directory,
and the stand-alone binary for running the frame labeler in isolation
has been retired