PTSD-363 Fixed relatively minor issue, where the lifetime of a caching IAllocator was too short.
The baz writing thread was holding on to a baz buffer after the point at which
the allocator that generated it (from the prelim-hq filter) was destroyed.

This issue was mostly benign, since it can only occur at the end of execution, and we "safely"
just leak the memory rather than trying ot access/use destroyed objects.  It also was ultimately
transient, since this only happens because we have a standalone thread handling baz io, rather
than properly integrating it with the computational graph.  Still, the issue was resolved by
updating some allocator references stores in the baz buffer to being full shared_ptrs, which
extends the lifetime as necessary.