PTSD-958 Fixed tbb issue requiring the intel module to be loaded.
This change was off topic even for PTSD-958, but there was an intended
fix for this problem under that issue.  That fix was incomplete,
because the change in pa-third-party that resolved the issue had
been merged but never built.

This change just moves to the newest pa-third-party artifact and
removes the intel module load statement that was hiding the issue.