PTSD-599 Added some missing plumbing for frame labeler configs
None of the frame labeler configs were actually plumbed through
to their intended usage sites.  That has been fixed for the
parameters relating to the subframe scorer for both the host
and the gpu version.  A stub has been placed in the gpu side
for receiving the roi parameters, so they are ready to be
used in the upcoming implementation