PTSD-599 Moved from a one pass to a two pass algorithm
The original approach to use the roi information during
the traceback step was fundamentally flawed.  Altering
the traceback does sort of have the same effect as
adjusting the loglikelihood with a "penalty" for precicely
one frame, however the loglikelihoods for each state in
subsequent frames won't quite be correct and could well
introduce errors.

Things now have gone back to the Sequel approach, where the
roi is computed upfront, and non-roi states apply a negative
infinity "penalty" to each non-baseline destination states