PTSD-1260: Added some constraints to the DME algorithm.
This is a bandaid more than anything, but a few constraints
were added to prevent various peices of the DME from underflowing
or overflowing.  Mostly the variances are constrained to be finite,
but there are a couple other things like the analog means being
positive during the model initialization.

These artificial constraints likely are masking deeper issues that
still need to be addressed