PTSD-1270 Added explicit step to release gpu memory after a batch is finished
This is to work around an issue where the trace saver had a performance
hiccough, which wasn't slow enough to cause a realtime violation, but
it was slow enough that the basecaller could analyze most/all the batches
before they could have their traces written.  Before this commit, this means
none of the gpu memory for that chunk was released until the end of the chunk,
meaning we had ~100 batches of data in flight rather than 3-5, and balooning
our gpu memory usage to the point where we died.

Now we explicitly release the GPU memory once the batch is finished processing.
It's definitely a rough edge of BatchData abstraction, that this kind of
coupling between separate system components is possible, but smoothing that
rough edge would require some extra design work