PTSD-964 Fixed some plumbing issues in the baseline autocorr computation
The m2lag value was kept around as a persistent "latent" data, and never
reset, which means it growed continuously until overflow.

The fix was to treat the m2lag as the rest of the autocorrelation values,
as members of the "local latent" struct, which are zero initialized each
chunk.  Probably would be good to just separate those out into a new
struct unrelated to "latent" data, but that's a larger scale change I'm
not currently prepared to take on