PTSD-1414 Put in experiemental retry mechanism for calls to cudaHostRegister calls
This is a complete shot-in-the-dark experiemental change.  To make sure that we actually
know if this succeeded, we'll still throw an exception in the "success" case to make
sure someone still reports it to us (since log messages without a crash will probably
just get lost in the void).

Once we get any hint of success we'll quickly update this so that a successful rety
does not terminate the application