PTSD-1525 Fixed various issues in the device hf metrics, and finally migrated things completely away from float2 and towards PBFloat2
The big one was that pkzvar was truncated from a single precision accumulation variable to a half precision pkzvar too early,
causing us to overflow.  Now the whole thing is done in single precision (though pkzvar is still stored as half precision
during the cart traversal)

The next biggest was handling an infinity that crops up when the baseline ia too close to one of the analogs, which messes
up the pkzvar normalization.  Effectively infinite values are now capped to a finite "large" value