PTSD-1525 Fixed issue where pkmid was serialized as nan instead of 0
There was some confusion, since the original Sequel code this was based
on tended to flip flop between nan and 0 as magic values for pkmid metrics
when a given analog was not present.  This only worked previously because
of a separate bug, where we tried to store a nan in the metrics, but
accidentally stored a 0 instead.  This masked the missing conversion contained
in this patch, where a metric block with 0 frames in a given analog just
returns 0 for the pkmid