PTSD-944 Added a warmup phase that runs through the DeviceOnlyArray instances to make some of the lazy allocations less lazy
DeviecOnlyArray needs at least somewhat lazy allocations, because until all the DOA are constructed and `Partition`
is called on the stash, we don't know which objects live on the host and which fit on the GPU.  However if we let
it be too lazy then we'll end up having a bunch of synchronous cuda allocation function calls that end up
catastrophically slowing down the analysis of the first batch.

So as a quick workaround we just cycle through all the data once we call Partition, effectively getting the major
allocations out of the way while we're not yet time sensitive.  It's a bit of a hack, this should be functionality
built directly into the cache, but doing so would be a slightly larger refactor than I want to take on at this
moment in time