PTSD-944 Made the host allocations for UnifiedCudaArray unrealized unless necessary.
Previously, the UnifiedCudaArray *always* had an associated host allocation.  That
was a bit wasteful for data that was set explicitly to be populated on the gpu
and downloaded on the host, but were never passed to a piece of code that actually
downloaded it on the host.

Now we only have an initial allocation matching the activeOnHost_ flag, and any
missing allocation is provisioned just in time if necessary for a download.
This both saves us a little bit of host ram, for some data that theoretically
could be downloaded from the GPU but never actually is, but also prevents an
unecessary cudaMallocHost function call during the analysis of the first chunk,
which is something that has a significant performance impact

This did force us to jump through a couple hoops, say when generating an
"empty trace back" because we're short circuiting our computation since
the DME has not run.  In that case we now have to be careful to just allocate
the data on the host, since otherwise now we'd needlessly allocate it on the
gpu and then download the junk data to the host